Renamed eslintrc.json to .eslintrc.json as per Piazaa

+++ app/app/src/main/java/com/example/cpen321app/MainActivity.kt
@@ -429,7 +429,7 @@ class MainActivity : AppCompatActivity() {
                             name = taskJson.getString("name"),
                             start = taskJson.getString("start"),
                             end = taskJson.getString("end"),
-                            duration = taskJson.getDouble("duration"),
+                            duration = taskJson.getInt("duration"),
                             location_lat = taskJson.getDouble("location_lat"),
                             location_lng = taskJson.getDouble("location_lng"),
                             priority = taskJson.getInt("priority"),
